# docs/swagger.yaml
basePath: /
definitions:
...
  sales.SalesDetail:
    properties:
      amount:
        type: number
      id:
        type: string
      item_code:
        type: string
      item_description:
        type: string
      item_id:
        type: string
      price:
        type: number
      quantity:
        type: integer
      sales_header_id:
        type: string
    required:
    - amount
    - id
    - item_code
    - item_description
    - item_id
    - price
    - quantity
    - sales_header_id
    type: object
  sales.SalesDetailRequest:
    properties:
      amount:
        type: number
      item_id:
        type: string
      price:
        type: number
      quantity:
        type: integer
      sales_header_id:
        type: string
    required:
    - amount
    - item_id
    - price
    - quantity
    - sales_header_id
    type: object
  sales.SalesHeader:
    properties:
      code:
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      id:
        type: string
      sent:
        type: boolean
    required:
    - code
    - created_at
    - customer_name
    - id
    - sent
    type: object
  sales.SalesHeaderRequest:
    properties:
      code:
        type: string
      customer_name:
        type: string
    required:
    - code
    - customer_name
    type: object
...
host: localhost:8125
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API per gestionar compres, ventes i stocks bàsica
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FRDY API
  version: "1.0"
paths:
...
  /api/sales/customers/{customerName}:
    get:
      consumes:
      - application/json
      description: Retrieve sales information by customer name (Protected route)
      parameters:
      - description: Customer Name
        in: path
        name: customerName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sales.SalesHeader'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sales by customer name
      tags:
      - sales-queries
  /api/sales/details:
    post:
      consumes:
      - application/json
      description: Create a new sales detail with the provided information (Protected
        route)
      parameters:
      - description: Sales detail data
        in: body
        name: detail
        required: true
        schema:
          $ref: '#/definitions/sales.SalesDetailRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sales.SalesDetail'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new sales detail
      tags:
      - sales-details
  /api/sales/details/{headerID}:
    get:
      consumes:
      - application/json
      description: Retrieve all sales details for a specific header ID (Protected
        route)
      parameters:
      - description: Sales Header ID
        in: path
        name: headerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sales.SalesDetail'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sales details by header ID
      tags:
      - sales-details
  /api/sales/details/{id}:
    delete:
      consumes:
      - application/json
      description: Delete sales detail information by ID (Protected route)
      parameters:
      - description: Sales Detail ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete sales detail by ID
      tags:
      - sales-details
    put:
      consumes:
      - application/json
      description: Update sales detail information by ID (Protected route)
      parameters:
      - description: Sales Detail ID
        in: path
        name: id
        required: true
        type: string
      - description: Sales detail update data
        in: body
        name: detail
        required: true
        schema:
          $ref: '#/definitions/sales.SalesDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sales.SalesDetail'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a sales detail
      tags:
      - sales-details
  /api/sales/headers:
    get:
      consumes:
      - application/json
      description: Retrieve all sales headers (Protected route)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sales.SalesHeader'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all sales
      tags:
      - sales-headers
    post:
      consumes:
      - application/json
      description: Create a new sales header with the provided information (Protected
        route)
      parameters:
      - description: Sales header data
        in: body
        name: header
        required: true
        schema:
          $ref: '#/definitions/sales.SalesHeaderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sales.SalesHeader'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new sales header
      tags:
      - sales-headers
  /api/sales/headers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete sales information by header ID (Protected route)
      parameters:
      - description: Sales Header ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete sales by header ID
      tags:
      - sales-headers
    get:
      consumes:
      - application/json
      description: Retrieve sales information by header ID (Protected route)
      parameters:
      - description: Sales Header ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sales.SalesHeader'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sales by header ID
      tags:
      - sales-headers
    put:
      consumes:
      - application/json
      description: Update sales header information by ID (Protected route)
      parameters:
      - description: Sales Header ID
        in: path
        name: id
        required: true
        type: string
      - description: Sales header update data
        in: body
        name: header
        required: true
        schema:
          $ref: '#/definitions/sales.SalesHeaderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sales.SalesHeader'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a sales header
      tags:
      - sales-headers
  /api/sales/headers/code/{code}:
    get:
      consumes:
      - application/json
      description: Retrieve sales information by header code (Protected route)
      parameters:
      - description: Sales Header Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sales.SalesHeader'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sales by header code
      tags:
      - sales-headers
  /api/sales/headers/send/{id}:
    post:
      consumes:
      - application/json
      description: Send/confirm a sales header by ID (Protected route)
      parameters:
      - description: Sales Header ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sales.SalesHeader'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send sales header
      tags:
      - sales-headers
  /api/sales/items/{itemCode}:
    get:
      consumes:
      - application/json
      description: Retrieve sales information by item code (Protected route)
      parameters:
      - description: Item Code
        in: path
        name: itemCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sales.SalesHeader'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sales by item code
      tags:
      - sales-queries
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"